+++ NEWS.md
@@ -23,6 +23,24 @@
 - Optimize frame to save memory
 - Fix bugs
 
+
 #### Version 0.1.5 (20191013)
 --------------
 - Add CcsAnnotate Function.
+
+#### Version 0.1.6 (20191101)
+--------------
+- Change the default output directory in CCS prediction.
+
+#### Version 0.1.61 (20191102)
+--------------
+- Fix some known bugs.
+
+#### Version 0.1.62 (20191109)
+--------------
+- Change default thread to 1 in CcsAnnotation.
+- Add some dependences packages
+
+#### Verison 0.1.63 (20200403)
+--------------
+o Package for submission